chalk terminal errors
fix callback log error for missing directory and file

+++ packages/netlify-cms-backend-fs/scripts/fs/fs-api.js
@@ -1,25 +1,38 @@
 const fs = require('fs');
 const path = require('path');
+const chalk = require('chalk');
+const log = console.log;
 
 const projectRoot = path.join(process.cwd());
-console.log(`Root path is ${ projectRoot }`);
+log(chalk.green(`Root path is ${ projectRoot }`));
 
 const siteRoot = {
   dir: path.join(projectRoot, "example")
 };
 const setPath = (relPath) => {
   siteRoot.dir = path.join(projectRoot, relPath);
-  console.log(`Site path is ${ siteRoot.dir }`);
+  log(chalk.green(`Site path is ${ siteRoot.dir }`));
 };
+
+const logError = (message) => {
+  log(chalk.red(message));
+  return new Error(message);
+}
+
+const notCallback = (cb) => {
+  return (typeof cb !== "function")
+}
 
 module.exports = {
   site: { setPath },
   files: (dirname) => {
     const name = "Files";
     const read = (cb) => {
-      if (!cb) throw new Error(`Invalid call to ${name}.read - requires a callback function(content)`);
+      if (notCallback(cb)) throw logError(`${name}.read: missing callback(content)`);
       const thispath = path.join(siteRoot.dir, dirname);
-      const files = fs.existsSync(thispath) ? fs.readdirSync(thispath) : [];
+      const dirExists = fs.existsSync(thispath);
+      if (!dirExists) log(chalk.yellow(`Warning: Directory missing ${thispath}`));
+      const files =  dirExists ? fs.readdirSync(thispath) : [];
       const filelist = [];
       files.forEach(function(element) {
         const filePath = path.join(thispath, element);
@@ -35,16 +48,21 @@ module.exports = {
   file: (id) => {
     const name = "File";
     const thisfile = path.join(siteRoot.dir, id);
-    let stats;
-    try {
-      stats = fs.statSync(thisfile);
-    } catch (err) {
-      stats = { };
+
+    const readStats = (path) => {
+      let stats;
+      try {
+        stats = fs.statSync(thisfile);
+      } catch (err) {
+        stats = {};
+      }
+      return stats;
     }
 
     /* GET-Read an existing file */
     const read = (cb) => {
-      if (!cb) throw new Error(`Invalid call to ${name}.read - requires a callback function(content)`);
+      if (notCallback(cb)) throw logError(`${name}.read: missing callback(content)`);
+      const stats = readStats(thisfile);
       if (typeof stats.isFile === "function" && stats.isFile()) {
         fs.readFile(thisfile, 'utf8', (err, data) => {
           if (err) {
@@ -54,11 +72,12 @@ module.exports = {
           }
         });
       } else {
-        throw new Error(`Invalid call to ${name}.read - object path is not a file(${thisfile})`);
+        cb({ error: logError(`${name}.read: - not a file(${thisfile})`) });
       }
     };
     /* POST-Create a NEW file, ERROR if exists */
     const create = (body, cb) => {
+      if (notCallback(cb)) throw logError(`${name}.create: missing callback(content)`);
       if (fs.existsSync(thisfile)) throw new Error(`Invalid call to ${name}.create - File does exists(${thisfile})`);
       fs.writeFile(thisfile, body.content, { encoding: body.encoding, flag: 'wx' }, (err) => {
         if (err) {
@@ -70,6 +89,7 @@ module.exports = {
     };
     /* PUT-Update an existing file */
     const update = (body, cb) => {
+      if (notCallback(cb)) throw logError(`${name}.update: missing callback(content)`);
       if (!fs.existsSync(thisfile)) throw new Error(`Invalid call to ${name}.update - File does not exist(${thisfile})`);
       fs.writeFile(thisfile, body.content, { encoding: body.encoding, flag: 'w' }, (err) => {
         if (err) {
@@ -81,6 +101,7 @@ module.exports = {
     };
     /* DELETE an existing file */
     const del = (cb) => {
+      if (notCallback(cb)) throw logError(`${name}.del: missing callback(content)`);
       fs.unlink(thisfile, (err) => {
         if (err) {
           cb({ error: err });
@@ -89,6 +110,6 @@ module.exports = {
         }
       });
     };
-    return { read, create, update, del, stats };
+    return { read, create, update, del };
   },
 };
